Renamed projects to watchlist
