+++ frontend/timeline.html
@@ -31,7 +31,7 @@
     <script src="/js/helpers/metric-boxes.js" defer></script>
     <script src="/js/helpers/converters.js" defer></script>
     <script src="/js/helpers/charts.js" defer></script>
-    <script src="/js/scenario-runner-timeline.js" defer></script>
+    <script src="/js/timeline.js" defer></script>
     <link rel="stylesheet" type="text/css" class="ui" href="/dist/css/semantic_reduced.min.css">
     <link rel="stylesheet" type="text/css" href="/css/green-coding.css">
 </head>
