+++ frontend/watchlist.html
@@ -1,0 +1,48 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <!-- Standard Meta -->
+    <meta charset="utf-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
+    <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1">
+    <link rel="shortcut icon" href="/images/favicon.ico">
+    <link rel="icon" href="/images/favicon.ico">
+    <link rel="icon" href="/images/favicon-16x16.png" sizes="16x16">
+    <link rel="icon" href="/images/favicon-32x32.png" sizes="32x32">
+    <!-- Site Properties -->
+    <title>Green Metrics Tool - ScenarioRunner Watchlist</title>
+    <meta name="referrer" content="no-referrer-when-downgrade" />
+    <meta name="description" content="Measuring the energy and carbon footprint of your software" />
+    <script src="/dist/js/jquery.min.js" defer></script>
+    <script src="/dist/js/accordion.min.js" defer></script>
+    <script src="/dist/js/transition.min.js" defer></script>
+    <script src="/dist/js/toast.min.js" defer></script>
+    <script src="/js/helpers/config.js" defer></script>
+    <script src="/js/helpers/main.js" defer></script>
+    <script src="/js/watchlist.js" defer></script>
+
+    <link rel="stylesheet" type="text/css" class="ui" href="/dist/css/semantic_reduced.min.css">
+    <link rel="stylesheet" type="text/css" href="/css/green-coding.css">
+</head>
+<body class="preload">
+    <gmt-menu></gmt-menu>
+    <div class="main ui container" id="main">
+        <h1 class="ui header float left">
+            <a id="menu-toggle" class="opened"><i class="bars bordered inverted left icon openend"></i></a>
+            ScenarioRunner - Watchlist
+        </h1>
+        <div class="ui full-width-card segment card">
+            <div class="content">
+                <div class="header">ScenarioRunner Watchlist Overview</div>
+                <div class="description">
+                    <i class="close icon"></i>
+                    <p>This page shows you all repositories that are monitored via ScenarioRunner at the moment</p>
+                    <p>The score cards show you the trends on specific key metrics and also the monitoring interval / last date measured.</p>
+                    <p>When you click in the badges you will get to the timeline view, where you can see all badges and more charts and metrics.</p>
+                </div>
+            </div>
+        </div>
+        <div id="scenario-runner-watchlist"class="ui cards"></div>
+    </div>
+</body>
+</html>
